# Constructing a ResourceIdentifier no longer allocates 5 strings