--- a/.gitignore
+++ b/.gitignore
@@ -1,28 +1,54 @@
-# Gradle
-*.class
-*.jar
-.gradle
-build/
-
 # Eclipse
+*.class
+.project
+.gradle
 .classpath
 .checkstyle
-.factorypath
-.project
 .settings
-bin/
-generated/
+.node
+/build/
+*/build/
+examples/*/build/
+extras/*/build/
+bin
+.factorypath
 
-# Idea
+# SLSv2
+!*/service/bin
+**/var/log
+**/var/data
+
+# Mobile Tools for Java (J2ME)
+.mtj.tmp/
+
+# Package Files #
+*.jar
+*.war
+*.ear
+
+# virtual machine crash logs, see http://www.java.com/en/download/help/error_hotspot.xml
+hs_err_pid*
+
+# Windows git artifact
+fileeditor
+
+# Mac specific files
+.DS_Store
+
+# IntelliJ
 *.iml
 *.ipr
 *.iws
 .idea/
 out/
+
+# Eclipse/IntelliJ APT
 generated_src/
+generated_testSrc/
+generated/
 
-# Codegen
-.generated
+# Blueprint theme
+__init__.pyc
 
-# Mac
-.DS_Store
+# Editor files
+*~