--- a/.gitignore
+++ b/.gitignore
@@ -1,54 +1,30 @@
-# Eclipse
+# Gradle
 *.class
-.project
+*.jar
 .gradle
+build/
+
+# Eclipse
 .classpath
 .checkstyle
-.settings
-.node
-/build/
-*/build/
-examples/*/build/
-extras/*/build/
-bin
 .factorypath
+.project
+.settings
+bin/
+generated/
 
-# SLSv2
-!*/service/bin
-**/var/log
-**/var/data
-
-# Mobile Tools for Java (J2ME)
-.mtj.tmp/
-
-# Package Files #
-*.jar
-*.war
-*.ear
-
-# virtual machine crash logs, see http://www.java.com/en/download/help/error_hotspot.xml
-hs_err_pid*
-
-# Windows git artifact
-fileeditor
-
-# Mac specific files
-.DS_Store
-
-# IntelliJ
+# Idea
 *.iml
 *.ipr
 *.iws
 .idea/
 out/
-
-# Eclipse/IntelliJ APT
 generated_src/
 generated_testSrc/
 generated/
 
-# Blueprint theme
-__init__.pyc
+# Codegen
+.generated
 
-# Editor files
-*~
+# Mac
+.DS_Store