--- a/build.gradle
+++ b/build.gradle
@@ -12,6 +12,7 @@ buildscript {
         classpath 'com.palantir.baseline:gradle-baseline-java:3.92.0'
         classpath 'com.palantir.gradle.consistentversions:gradle-consistent-versions:1.28.0'
         classpath 'com.palantir.gradle.gitversion:gradle-git-version:0.12.3'
+        classpath 'me.champeau.jmh:jmh-gradle-plugin:0.6.5'
     }
 }
 allprojects {
@@ -23,6 +24,7 @@ apply plugin: 'java-library'
 apply plugin: 'com.palantir.baseline'
 apply plugin: 'com.palantir.git-version'
 apply plugin: 'com.palantir.consistent-versions'
+apply plugin: 'me.champeau.jmh'
 
 group 'com.palantir.ri'
 version gitVersion()
@@ -41,4 +43,20 @@ dependencies {
 
     testImplementation 'junit:junit'
     testImplementation 'com.fasterxml.jackson.core:jackson-databind'
+
+    annotationProcessor 'org.openjdk.jmh:jmh-generator-annprocess'
+    compileOnly 'org.openjdk.jmh:jmh-generator-annprocess'
+    jmh 'org.openjdk.jmh:jmh-core'
+    // jmh 'org.apache.logging.log4j:log4j-slf4j-impl'
+}
+
+jmh {
+    // Use profilers to collect additional data. Supported profilers:
+    // [cl, comp, gc, stack, perf, perfnorm, perfasm, xperf, xperfasm, hs_cl, hs_comp, hs_gc, hs_rt, hs_thr]
+    profilers = ['gc']
+}
+
+tasks.jmhCompileGeneratedClasses {
+    options.annotationProcessorPath = configurations.errorprone
+    options.errorprone.enabled = false
 }