--- a/build.gradle
+++ b/build.gradle
@@ -12,6 +12,7 @@ buildscript {
         classpath 'com.palantir.baseline:gradle-baseline-java:3.92.0'
         classpath 'com.palantir.gradle.consistentversions:gradle-consistent-versions:1.28.0'
         classpath 'com.palantir.gradle.gitversion:gradle-git-version:0.12.3'
+        classpath 'gradle.plugin.org.inferred:gradle-processors:3.3.0'
         classpath 'me.champeau.jmh:jmh-gradle-plugin:0.6.5'
     }
 }
@@ -24,6 +25,7 @@ apply plugin: 'java-library'
 apply plugin: 'com.palantir.baseline'
 apply plugin: 'com.palantir.git-version'
 apply plugin: 'com.palantir.consistent-versions'
+apply plugin: 'org.inferred.processors'
 apply plugin: 'me.champeau.jmh'
 
 group 'com.palantir.ri'