--- a/build.gradle
+++ b/build.gradle
@@ -52,8 +52,6 @@ dependencies {
 }
 
 jmh {
-    // Use profilers to collect additional data. Supported profilers:
-    // [cl, comp, gc, stack, perf, perfnorm, perfasm, xperf, xperfasm, hs_cl, hs_comp, hs_gc, hs_rt, hs_thr]
     profilers = ['gc']
 }