--- a/versions.lock
+++ b/versions.lock
@@ -2,8 +2,7 @@
 com.fasterxml.jackson.core:jackson-annotations:2.11.3 (2 constraints: bf170e71)
 net.sf.jopt-simple:jopt-simple:4.6 (1 constraints: 610a91b7)
 org.apache.commons:commons-math3:3.2 (1 constraints: 5c0a8ab7)
-org.openjdk.jmh:jmh-core:1.32 (1 constraints: 1411a7be)
-org.openjdk.jmh:jmh-generator-annprocess:1.32 (1 constraints: da04f730)
+org.openjdk.jmh:jmh-core:1.32 (1 constraints: da04f730)
 
 [Test dependencies]
 com.fasterxml.jackson.core:jackson-core:2.11.3 (1 constraints: 87123221)