--- a/versions.props
+++ b/versions.props
@@ -1,2 +1,3 @@
 com.fasterxml.jackson.core:* = 2.11.3
 junit:junit = 4.13.2
+org.openjdk.jmh:* = 1.32