## Before this PR (544 bytes/op)

Deserializing a ResourceIdentifier from JSON involves running a regex on the entire thing, then pulling out all the matches into 4 separate strings (which allocates on the heap) and then immediately joining them back together to allocate a fifth string which was equivalent to the original string that we read from JSON.

This convoluted splitting and then re-joining just creates work to keep the GC algorithm warm, and is actually responsible for a surprising amount of allocation (even if it gets GC'd by the new collector super quickly).

![image](https://user-images.githubusercontent.com/3473798/122286797-26958600-cee8-11eb-9119-894698e37f6f.png)

```
Benchmark                                                                    (input)  Mode  Cnt     Score      Error   Units
ResourceIdentifierBenchmark.fromString                                   NO_INSTANCE  avgt    3  1117.971 ± 7036.423   ns/op
ResourceIdentifierBenchmark.fromString:·gc.alloc.rate                    NO_INSTANCE  avgt    3  1703.296 ± 8968.591  MB/sec
ResourceIdentifierBenchmark.fromString:·gc.alloc.rate.norm               NO_INSTANCE  avgt    3   544.039 ±    0.019    B/op
ResourceIdentifierBenchmark.fromString:·gc.churn.G1_Eden_Space           NO_INSTANCE  avgt    3  1715.801 ± 8858.257  MB/sec
ResourceIdentifierBenchmark.fromString:·gc.churn.G1_Eden_Space.norm      NO_INSTANCE  avgt    3   548.811 ±  201.537    B/op
ResourceIdentifierBenchmark.fromString:·gc.churn.G1_Survivor_Space       NO_INSTANCE  avgt    3     0.006 ±    0.097  MB/sec
ResourceIdentifierBenchmark.fromString:·gc.churn.G1_Survivor_Space.norm  NO_INSTANCE  avgt    3     0.002 ±    0.023    B/op
ResourceIdentifierBenchmark.fromString:·gc.count                         NO_INSTANCE  avgt    3    49.000             counts
ResourceIdentifierBenchmark.fromString:·gc.time                          NO_INSTANCE  avgt    3    31.000                 ms
ResourceIdentifierBenchmark.fromString                                          UUID  avgt    3   856.617 ±  139.048   ns/op
ResourceIdentifierBenchmark.fromString:·gc.alloc.rate                           UUID  avgt    3  2227.241 ±  356.649  MB/sec
ResourceIdentifierBenchmark.fromString:·gc.alloc.rate.norm                      UUID  avgt    3   584.040 ±    0.035    B/op
ResourceIdentifierBenchmark.fromString:·gc.churn.G1_Eden_Space                  UUID  avgt    3  2238.919 ± 1109.428  MB/sec
ResourceIdentifierBenchmark.fromString:·gc.churn.G1_Eden_Space.norm             UUID  avgt    3   587.227 ±  386.892    B/op
ResourceIdentifierBenchmark.fromString:·gc.churn.G1_Survivor_Space              UUID  avgt    3     0.007 ±    0.032  MB/sec
ResourceIdentifierBenchmark.fromString:·gc.churn.G1_Survivor_Space.norm         UUID  avgt    3     0.002 ±    0.008    B/op
ResourceIdentifierBenchmark.fromString:·gc.count                                UUID  avgt    3    64.000             counts
ResourceIdentifierBenchmark.fromString:·gc.time                                 UUID  avgt    3    37.000                 ms
```


## After this PR (248 bytes/op)

==COMMIT_MSG==
Deserializing a ResourceIdentifier now causes zero heap-allocated strings (previously it allocated 5).
==COMMIT_MSG==


```

Benchmark                                                                    (input)  Mode  Cnt     Score      Error   Units
ResourceIdentifierBenchmark.fromString                                   NO_INSTANCE  avgt    3   694.585 ±  592.517   ns/op
ResourceIdentifierBenchmark.fromString:·gc.alloc.rate                    NO_INSTANCE  avgt    3  1167.447 ± 1033.632  MB/sec
ResourceIdentifierBenchmark.fromString:·gc.alloc.rate.norm               NO_INSTANCE  avgt    3   248.022 ±    0.005    B/op
ResourceIdentifierBenchmark.fromString:·gc.churn.G1_Eden_Space           NO_INSTANCE  avgt    3  1196.048 ± 1620.917  MB/sec
ResourceIdentifierBenchmark.fromString:·gc.churn.G1_Eden_Space.norm      NO_INSTANCE  avgt    3   253.903 ±  138.877    B/op
ResourceIdentifierBenchmark.fromString:·gc.churn.G1_Survivor_Space       NO_INSTANCE  avgt    3     0.004 ±    0.012  MB/sec
ResourceIdentifierBenchmark.fromString:·gc.churn.G1_Survivor_Space.norm  NO_INSTANCE  avgt    3     0.001 ±    0.002    B/op
ResourceIdentifierBenchmark.fromString:·gc.count                         NO_INSTANCE  avgt    3    42.000             counts
ResourceIdentifierBenchmark.fromString:·gc.time                          NO_INSTANCE  avgt    3    24.000                 ms
ResourceIdentifierBenchmark.fromString                                          UUID  avgt    3   825.575 ±  332.685   ns/op
ResourceIdentifierBenchmark.fromString:·gc.alloc.rate                           UUID  avgt    3   982.036 ±  395.425  MB/sec
ResourceIdentifierBenchmark.fromString:·gc.alloc.rate.norm                      UUID  avgt    3   248.022 ±    0.010    B/op
ResourceIdentifierBenchmark.fromString:·gc.churn.G1_Eden_Space                  UUID  avgt    3   998.716 ±  911.556  MB/sec
ResourceIdentifierBenchmark.fromString:·gc.churn.G1_Eden_Space.norm             UUID  avgt    3   252.394 ±  287.709    B/op
ResourceIdentifierBenchmark.fromString:·gc.churn.G1_Survivor_Space              UUID  avgt    3     0.004 ±    0.028  MB/sec
ResourceIdentifierBenchmark.fromString:·gc.churn.G1_Survivor_Space.norm         UUID  avgt    3     0.001 ±    0.007    B/op
ResourceIdentifierBenchmark.fromString:·gc.count                                UUID  avgt    3    35.000             counts
ResourceIdentifierBenchmark.fromString:·gc.time                                 UUID  avgt    3    21.000                 ms
```