@svc-autorelease 2021-06-16
Released [1.3.0](https://github.com/palantir/resource-identifier/releases/tag/1.3.0)